requirements.txt
# ─── Web Framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# ─── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.35
aiosqlite==0.20.0

# ─── Validation ───────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ─── LLM / Agentic AI ─────────────────────────────────────────
# tenacity<9.0.0 is required by langchain 0.2.x — pinned to 8.5.0
openai==1.45.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langgraph==0.2.28
tenacity==8.5.0

# ─── RAG / Vector Store ───────────────────────────────────────
chromadb==0.5.12
sentence-transformers==3.1.1

# ─── Computer Vision / ML ─────────────────────────────────────
# CPU-only torch — smaller download, sufficient for ResNet18 inference
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
scikit-learn==1.5.2

# ─── Forecasting ──────────────────────────────────────────────
prophet==1.1.5
pandas==2.2.3
numpy==1.26.4

# ─── HTTP / Async ─────────────────────────────────────────────
httpx==0.27.2
requests==2.32.3

# ─── Dashboard ────────────────────────────────────────────────
streamlit==1.38.0
plotly==5.24.1

# ─── AIOps / Monitoring ───────────────────────────────────────
psutil==6.0.0
boto3==1.35.23

# ─── Utilities ────────────────────────────────────────────────
loguru==0.7.2
aiofiles==24.1.0
schedule==1.2.2

# ─── Testing ──────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
